ticketai-backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
PyMySQL==1.1.1
django-ratelimit==4.1.0
argon2-cffi==23.1.0
httpx==0.27.0
django-filter==24.2
python-decouple==3.8
pyotp
jsonschema